[MDB IGNORE] [IDB IGNORE] upstream merge 05/12/2026

+++ code/game/objects/structures/water_structures/sink.dm
@@ -19,7 +19,7 @@
 	///Does the sink have a water recycler to recollect its water supply?
 	var/has_water_reclaimer = TRUE
 	///Units of water to reclaim per second
-	var/reclaim_rate = 0.5
+	var/reclaim_rate = 50 // DARKPACK EDIT CHANGE - ORIGINAL: var/reclaim_rate = 0.5
 
 MAPPING_DIRECTIONAL_HELPERS(/obj/structure/sink, (-14))
 
@@ -29,7 +29,7 @@ MAPPING_DIRECTIONAL_HELPERS(/obj/structure/sink, (-14))
 	create_reagents(capacity, NO_REACT)
 	if(has_water_reclaimer)
 		reagents.add_reagent(dispensedreagent, capacity)
-	AddComponent(/datum/component/plumbing/simple_demand/extended)
+	// AddComponent(/datum/component/plumbing/simple_demand/extended) // DARKPACK EDIT REMOVAL
 
 	register_context()
 
@@ -285,6 +285,7 @@ MAPPING_DIRECTIONAL_HELPERS(/obj/structure/sink/basin, (0))
 	icon = 'icons/obj/watercloset.dmi'
 	icon_state = "sink_frame"
 	desc = "A sink frame, that needs a water recycler to finish construction."
+	wall_external = TRUE
 	result_path = /obj/structure/sink/greyscale
 	material_flags = MATERIAL_EFFECTS | MATERIAL_ADD_PREFIX | MATERIAL_COLOR | MATERIAL_AFFECT_STATISTICS
 	pixel_shift = 16
@@ -306,10 +307,11 @@ MAPPING_DIRECTIONAL_HELPERS(/obj/structure/sink/basin, (0))
 
 /obj/item/wallframe/sinkframe/item_interaction(mob/living/user, obj/item/tool, list/modifiers)
 	. = NONE
-	if(istype(tool, /obj/item/stock_parts/water_recycler))
+	if(istype(tool, /obj/item/stock_parts/water_recycler) && result_path == /obj/structure/sink/greyscale)
 		qdel(tool)
 		result_path = /obj/structure/sink/greyscale/filled
 		playsound(src, 'sound/machines/click.ogg', 20, TRUE)
+		balloon_alert(user, "water recycler installed!")
 		return ITEM_INTERACT_SUCCESS
 
 /obj/item/wallframe/sinkframe/after_attach(obj/structure/sink/greyscale/attached_to)
