+++ modular_darkpack/modules/ritual_abyss_mysticism/code/rituals/artifact_identification.dm
@@ -6,8 +6,12 @@
 	cost = 1
 
 /obj/ritual_rune/abyss/identification/complete()
-	for(var/obj/item/vtm_artifact/VA in loc)
+	. = ..()
+	for(var/obj/item/occult_artifact/VA in loc)
+		var/mob/living/carbon/human/identifier = usr
+		if(VA.identified)
+			to_chat(identifier, span_warning("You have already identified this artifact."))
+			return
 		VA.identify()
-		playsound(loc, 'sound/effects/magic/voidblink.ogg', 50, FALSE)
 		qdel(src)
 		return
